backend/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
google-generativeai>=0.8.0
anthropic>=0.39.0
openai>=1.30.0
together>=1.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
requests>=2.31.0
firebase-admin>=6.4.0
numpy>=1.26.0
python-multipart>=0.0.9
aiofiles>=23.2.1
